+++ source/extensions/isaacsim.replicator.writers/python/tests/test_pose_writer.py
@@ -16,25 +16,26 @@
 import json
 import os
 
-import carb
 import numpy as np
 import omni.kit
 import omni.replicator.core as rep
 import omni.usd
+from isaacsim.test.utils.file_validation import validate_folder_contents
+from isaacsim.test.utils.image_comparison import compare_images_in_directories
 
 
-def compare_json_dicts_recursive(data1, data2, path=""):
+def compare_nested_structures_with_tolerance(data1, data2, path="", rtol=1e-5, atol=1e-5):
     if isinstance(data1, (list, tuple)) and isinstance(data2, (list, tuple)):
         if len(data1) != len(data2):
             return f"Length mismatch at {path}: {len(data1)} != {len(data2)}"
         if any(isinstance(x, float) for x in data1 + data2):
             try:
-                np.testing.assert_allclose(data1, data2, rtol=1e-5, atol=1e-5)
+                np.testing.assert_allclose(data1, data2, rtol=rtol, atol=atol)
             except AssertionError:
                 return f"Value mismatch at {path}: {data1} != {data2}"
         else:
             for i, (v1, v2) in enumerate(zip(data1, data2)):
-                error = compare_json_dicts_recursive(v1, v2, f"{path}[{i}]")
+                error = compare_nested_structures_with_tolerance(v1, v2, f"{path}[{i}]")
                 if error:
                     return error
     elif isinstance(data1, dict) and isinstance(data2, dict):
@@ -67,7 +68,9 @@ def compare_json_dicts_recursive(data1, data2, path=""):
                     except AssertionError:
                         return f"Quaternion mismatch at {path}.{key}: {data1[key]} != {data2[key]} (and not negative of each other)"
             else:
-                error = compare_json_dicts_recursive(data1[key], data2[key], f"{path}.{key}" if path else key)
+                error = compare_nested_structures_with_tolerance(
+                    data1[key], data2[key], f"{path}.{key}" if path else key
+                )
                 if error:
                     return error
     elif isinstance(data1, float) and isinstance(data2, float):
@@ -81,74 +84,134 @@ def compare_json_dicts_recursive(data1, data2, path=""):
 
 
 class TestPoseWriter(omni.kit.test.AsyncTestCase):
+
+    RGB_MEAN_DIFF_TOLERANCE = 10
+
     async def setUp(self):
-        pass
+        await omni.kit.app.get_app().next_update_async()
+        omni.usd.get_context().new_stage()
+        await omni.kit.app.get_app().next_update_async()
 
     async def tearDown(self):
-        pass
+        await omni.usd.get_context().close_stage_async()
+        await omni.kit.app.get_app().next_update_async()
+        # In some cases the test will end before the asset is loaded, in this case wait for assets to load
+        while omni.usd.get_context().get_stage_loading_status()[2] > 0:
+            await omni.kit.app.get_app().next_update_async()
 
     async def test_pose_writer(self):
         # Setup stage
         await omni.usd.get_context().new_stage_async()
-        rep.create.light(light_type="dome", intensity=500)
-        rep.create.cube(scale=(1, 1, 1), semantics=[("class", "cube")], position=(0, 0, 0), rotation=(0, 0, 0))
-        rep.create.cube(
-            scale=(1, 1, 1), semantics=[("class", "cube_rotated")], position=(-1, -1, 0), rotation=(0, 0, 45)
+        rep.functional.create.xform(name="World")
+        rep.functional.create.dome_light(intensity=500, parent="/World", name="DomeLight")
+        rep.functional.create.cube(
+            position=(0, 0, 0),
+            rotation=(0, 0, 0),
+            scale=(1, 1, 1),
+            name="Cube1",
+            parent="/World",
+            semantics=[("class", "cube")],
         )
-        rep.create.cube(
-            scale=(1.5, 0.5, 1), semantics=[("class", "cube_scaled")], position=(1, 1, 0), rotation=(75, 65, 0)
+        rep.functional.create.cube(
+            position=(-1, -1, 0),
+            rotation=(0, 0, 45),
+            scale=(1, 1, 1),
+            name="Cube2",
+            parent="/World",
+            semantics=[("class", "cube_rotated")],
+        )
+        rep.functional.create.cube(
+            position=(1, 1, 0),
+            rotation=(75, 65, 0),
+            scale=(1.5, 0.5, 1),
+            name="Cube3",
+            parent="/World",
+            semantics=[("class", "cube_scaled")],
         )
 
         # Create render products
-        cam1 = rep.create.camera(position=(0, 0, 10), look_at=(0, 0, 0), name="cam1")
-        cam2 = rep.create.camera(position=(-4, -4, 8), look_at=(0, 0, 0), name="cam2")
+        cam1 = rep.functional.create.camera(position=(0, 0, 10), look_at=(0, 0, 0), name="cam1")
+        cam2 = rep.functional.create.camera(position=(-4, -4, 8), look_at=(0, 0, 0), name="cam2")
         rp1 = rep.create.render_product(cam1, (512, 512), name="rp1")
         rp2 = rep.create.render_product(cam2, (512, 512), name="rp2")
         render_products = [rp1, rp2]
 
         # Setup writer
-        # out_dir = carb.tokens.get_tokens_interface().resolve("${temp}/test_pose_writer")
-        out_dir = os.path.join(os.getcwd(), "test_pose_writer")
+        out_dir = os.path.join(os.getcwd(), "_out_test_pose_writer")
+        backend = rep.backends.get("DiskBackend")
+        backend.initialize(output_dir=out_dir)
+        print(f"Output directory: {out_dir}")
         writer = rep.writers.get("PoseWriter")
-        writer.initialize(output_dir=out_dir, use_subfolders=True, write_debug_images=True)
+        writer.initialize(backend=backend, use_subfolders=True, write_debug_images=True)
         writer.attach(render_products)
 
         # Capture data
-        await rep.orchestrator.step_async()
+        await rep.orchestrator.step_async(rt_subframes=16)
         await rep.orchestrator.wait_until_complete_async()
 
-        # Golden json file paths
-        golden_dir = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data", "golden", "out_pose_writer")
-        golden_rp1_json = os.path.join(golden_dir, "rp1", "000000.json")
-        golden_rp2_json = os.path.join(golden_dir, "rp2", "000000.json")
-        self.assertTrue(os.path.exists(golden_rp1_json), f"Golden file {golden_rp1_json} does not exist")
-        self.assertTrue(os.path.exists(golden_rp2_json), f"Golden file {golden_rp2_json} does not exist")
-
-        # Output json file paths
-        rp1_json = os.path.join(out_dir, "rp1", "000000.json")
-        rp2_json = os.path.join(out_dir, "rp2", "000000.json")
-        self.assertTrue(os.path.exists(rp1_json), f"Output file {rp1_json} does not exist")
-        self.assertTrue(os.path.exists(rp2_json), f"Output file {rp2_json} does not exist")
-
-        # Compare the json golden output with the output for the two render products
-        with open(golden_rp1_json, "r") as f:
-            golden_rp1_data = json.load(f)
-        with open(rp1_json, "r") as f:
-            rp1_data = json.load(f)
-        error = compare_json_dicts_recursive(rp1_data, golden_rp1_data)
-        self.assertIsNone(error, f"'/rp1' comparison failed:\n{error}")
-
-        with open(golden_rp2_json, "r") as f:
-            golden_rp2_data = json.load(f)
-        with open(rp2_json, "r") as f:
-            rp2_data = json.load(f)
-        error = compare_json_dicts_recursive(rp2_data, golden_rp2_data)
-        self.assertIsNone(error, f"'/rp2' comparison failed:\n{error}")
-
-        # Clean up
+        # Clean up writer and render products
         writer.detach()
         for rp in render_products:
             rp.destroy()
             rp = None
         render_products = None
-        await omni.usd.get_context().close_stage_async()
+
+        # Test and golden directory paths
+        test_dir_rp1 = os.path.join(out_dir, "rp1")
+        test_dir_rp2 = os.path.join(out_dir, "rp2")
+        golden_dir = os.path.join(
+            os.path.dirname(os.path.realpath(__file__)), "data", "golden", "_out_test_pose_writer"
+        )
+        golden_dir_rp1 = os.path.join(golden_dir, "rp1")
+        golden_dir_rp2 = os.path.join(golden_dir, "rp2")
+
+        # Check if all the expected data has been written
+        folder_contents_success_rp1 = validate_folder_contents(path=test_dir_rp1, expected_counts={"json": 1, "png": 2})
+        self.assertTrue(folder_contents_success_rp1, f"Output directory contents validation failed for {test_dir_rp1}")
+        folder_contents_success_rp2 = validate_folder_contents(path=test_dir_rp2, expected_counts={"json": 1, "png": 2})
+        self.assertTrue(folder_contents_success_rp2, f"Output directory contents validation failed for {test_dir_rp2}")
+
+        # Compare all PNG images (RGB and overlay) with the golden images for rp1
+        result_rp1 = compare_images_in_directories(
+            golden_dir=golden_dir_rp1,
+            test_dir=test_dir_rp1,
+            path_pattern=r"\.png$",
+            allclose_rtol=None,  # Disable allclose for this test to rely only on mean tolerance
+            allclose_atol=None,
+            mean_tolerance=self.RGB_MEAN_DIFF_TOLERANCE,
+            print_all_stats=True,
+            print_per_file_results=True,
+        )
+        self.assertTrue(result_rp1["all_passed"], f"Image comparison failed for rp1: {result_rp1}")
+
+        # Compare all PNG images (RGB and overlay) with the golden images for rp2
+        result_rp2 = compare_images_in_directories(
+            golden_dir=golden_dir_rp2,
+            test_dir=test_dir_rp2,
+            path_pattern=r"\.png$",
+            allclose_rtol=None,  # Disable allclose for this test to rely only on mean tolerance
+            allclose_atol=None,
+            mean_tolerance=self.RGB_MEAN_DIFF_TOLERANCE,
+            print_all_stats=True,
+            print_per_file_results=True,
+        )
+        self.assertTrue(result_rp2["all_passed"], f"Image comparison failed for rp2: {result_rp2}")
+
+        # Compare the json golden output with the output for the two render products
+        golden_rp1_json = os.path.join(golden_dir_rp1, "000000.json")
+        test_rp1_json = os.path.join(test_dir_rp1, "000000.json")
+        with open(golden_rp1_json, "r") as f:
+            golden_rp1_data = json.load(f)
+        with open(test_rp1_json, "r") as f:
+            test_rp1_data = json.load(f)
+        error = compare_nested_structures_with_tolerance(test_rp1_data, golden_rp1_data, rtol=1e-5, atol=1e-5)
+        self.assertIsNone(error, f"'/rp1' comparison failed:\n{error}")
+
+        golden_rp2_json = os.path.join(golden_dir, "rp2", "000000.json")
+        test_rp2_json = os.path.join(test_dir_rp2, "000000.json")
+        with open(golden_rp2_json, "r") as f:
+            golden_rp2_data = json.load(f)
+        with open(test_rp2_json, "r") as f:
+            test_rp2_data = json.load(f)
+        error = compare_nested_structures_with_tolerance(test_rp2_data, golden_rp2_data, rtol=1e-5, atol=1e-5)
+        self.assertIsNone(error, f"'/rp2' comparison failed:\n{error}")